update styles JobPostPage, make one page
