--- apps/freelance/src/pages/JobPostPage/index.ts
@@ -1,2 +1,0 @@
-export { JobPostFirstPage } from './JobPostFirstPage';
-export { JobPostSecondPage } from './JobPostSecondPage';
