+++ apps/freelance/src/pages/JobPostPage/index.tsx
@@ -1,0 +1,11 @@
+import { Fragment } from 'react';
+import { AppBar, JobPostForm } from '@freelance/components';
+
+export default function JobPostPage() {
+  return (
+    <Fragment>
+      <AppBar />
+      <JobPostForm />
+    </Fragment>
+  );
+}
