+++ apps/freelance/src/translations/en.json
@@ -157,17 +157,16 @@
       "role_error": "Please make your choice"
     },
     "job_post_page": {
-      "page_name": "Post a job page",
-      "fill_this_form": "Please fill this form",
+      "form_title": "Please fill this form to post a job",
       "project_name_label": "Add a name for your project*",
       "about_label": "Tell more about your project*",
       "english_level_label": "English level*",
       "english_level_placeholder": "Required level",
       "job_category_label": "Please select job category*",
-      "skills_label": "What skills are required?",
+      "skills_label": "What skills are required?*",
       "skills_label_minimize": "Enter at least 3 skills",
       "skills_placeholder": "Start typing to search skills",
-      "duration": "Duration",
+      "duration": "Duration*",
       "duration_amount_placeholder": "Amount",
       "duration_placeholder": "Month/Week",
       "post_a_job": "Post a job"
