+++ libs/components/src/index.ts
@@ -14,8 +14,7 @@ export * from './lib/google-login-button/index';
 export * from './lib/input/input';
 export * from './lib/input/styles';
 export * from './lib/job-card';
-export * from './lib/JobPostFormFirstPage';
-export * from './lib/JobPostFormSecondPage';
+export * from './lib/JobPostForm';
 export * from './lib/jobs-app-bar/index';
 export * from './lib/login-form/login-form';
 export * from './lib/logo';
