+++ libs/components/src/lib/JobPostForm/styles.ts
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+export const FormWrapper = styled.div`
+  width: 500px;
+`;
+
+export const TitleWrapper = styled.div`
+  margin-top: 20px;
+  text-align: center;
+`;
+
+export const StyledErrorMessage = styled.p`
+  font-size: 14px;
+  font-weight: 500;
+  color: red;
+`;
+export const ButtonWrapper = styled.div`
+  margin-top: 15px;
+  display: flex;
+  justify-content: center;
+`;
