--- libs/components/src/lib/JobPostFormSecondPage/styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledErrorMessage = styled.p`
-  font-size: 14px;
-  font-weight: 500;
-  color: red;
-`;
